backend/requirements.txt
flask
flask-cors
flask-sqlalchemy
pyjwt
bcrypt
qrcode
pillow
python-dotenv
pytest
